Refactor doctor from Rails generator to rake task and remove irrelevant checks
- Convert Rails generator to proper namespaced rake task (rake react_on_rails:doctor)
- Remove irrelevant git status and hello_world checks that don't belong in a diagnostic tool
- Move SystemChecker from generators namespace to ReactOnRails module for better organization
- Update all documentation and tests to reflect the new rake task approach
- Maintain all core diagnostic functionality while improving usability

This makes the doctor command behave like other Rails diagnostic tools (e.g., rake routes)
rather than a code generation tool, which was inappropriate for its diagnostic purpose.
