+++ spec/lib/react_on_rails/doctor_spec.rb
@@ -1,0 +1,73 @@
+# frozen_string_literal: true
+
+require_relative "../../react_on_rails/spec_helper"
+require_relative "../../../lib/react_on_rails/doctor"
+
+RSpec.describe ReactOnRails::Doctor do
+  let(:doctor) { described_class.new(verbose: false, fix: false) }
+
+  describe "#initialize" do
+    it "initializes with default options" do
+      expect(doctor).to be_instance_of(described_class)
+    end
+
+    it "accepts verbose and fix options" do
+      verbose_doctor = described_class.new(verbose: true, fix: true)
+      expect(verbose_doctor).to be_instance_of(described_class)
+    end
+  end
+
+  describe "#run_diagnosis" do
+    before do
+      # Mock all output methods to avoid actual printing
+      allow(doctor).to receive(:puts)
+      allow(doctor).to receive(:exit)
+
+      # Mock file system interactions
+      allow(File).to receive_messages(exist?: false, directory?: false)
+      allow(doctor).to receive(:`).and_return("")
+
+      # Mock the checker to avoid actual system calls
+      checker = instance_double(ReactOnRails::SystemChecker)
+      allow(ReactOnRails::SystemChecker).to receive(:new).and_return(checker)
+      allow(checker).to receive_messages(
+        check_node_installation: true,
+        check_package_manager: true,
+        check_react_on_rails_packages: true,
+        check_shakapacker_configuration: true,
+        check_react_dependencies: true,
+        check_react_on_rails_initializer: true,
+        check_webpack_configuration: true,
+        add_success: true,
+        add_warning: true,
+        add_info: true,
+        errors?: false,
+        warnings?: false,
+        messages: []
+      )
+    end
+
+    it "runs diagnosis without errors" do
+      expect { doctor.run_diagnosis }.not_to raise_error
+    end
+
+    it "prints header" do
+      expect(doctor).to receive(:puts).with(/REACT ON RAILS DOCTOR/)
+      doctor.run_diagnosis
+    end
+
+    it "runs all check sections" do
+      checker = doctor.instance_variable_get(:@checker)
+
+      expect(checker).to receive(:check_node_installation)
+      expect(checker).to receive(:check_package_manager)
+      expect(checker).to receive(:check_react_on_rails_packages)
+      expect(checker).to receive(:check_shakapacker_configuration)
+      expect(checker).to receive(:check_react_dependencies)
+      expect(checker).to receive(:check_react_on_rails_initializer)
+      expect(checker).to receive(:check_webpack_configuration)
+
+      doctor.run_diagnosis
+    end
+  end
+end
